% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chiefdom}
\alias{chiefdom}
\title{Sierra Leone administrative area level 3 - chiefdom}
\format{
A SpatialPolygonsDataFrame with 167 features:
\describe{
  \item{\code{OBJECTID}}{Object Identifier}
  \item{\code{admin3Name}}{Name of administrative area level 3}
  \item{\code{admin3Pcod}}{Postal code of administrative area level 3}
  \item{\code{admin3RefN}}{Reference of administrative area level 3}
  \item{\code{admin3AltN}}{Alternative name of administrative area level 3}
  \item{\code{admin3Al_1}}{Alternative name of administrative area level 3}
  \item{\code{admin2Name}}{Name of administrative area level 2}
  \item{\code{admin2Pcod}}{Postal code of administrative area level 2}
  \item{\code{admin2RefN}}{Reference of administrative area level 2}
  \item{\code{admin2AltN}}{Alternative name of administrative area level 2}
  \item{\code{admin2Al_1}}{Alternative name of administrative area level 2}
  \item{\code{admin1Name}}{Name of administrative area level 1}
  \item{\code{admin1Pcod}}{Postal code of administrative area level 1}
  \item{\code{admin1RefN}}{Reference of administrative area level 1}
  \item{\code{admin1AltN}}{Alternative name of administrative area level 1}
  \item{\code{admin1Al_1}}{Alternative name of administrative area level 1}
  \item{\code{admin0Name}}{Name of administrative area level 0}
  \item{\code{admin0Pcod}}{Postal code of administrative area level 0}
  \item{\code{date}}{Date}
  \item{\code{validOn}}{Valid on}
  \item{\code{validTo}}{Valid to}
  \item{\code{Shape_Leng}}{Shape length}
  \item{\code{Shape_Area}}{Shape area}
}
}
\source{
\url{https://data.humdata.org/dataset/sierra-leone-all-ad-min-level-boundaries}
}
\usage{
chiefdom
}
\description{
Sierra Leone administrative area level 3 - chiefdom
}
\examples{
chiefdom

}
\keyword{datasets}
